fix scala warnings

+++ spark/src/main/scala/org/apache/comet/CometExecIterator.scala
@@ -86,7 +86,7 @@ class CometExecIterator(
   private val nativeLib = new Native()
   private val nativeUtil = new NativeUtil()
   private val taskAttemptId = TaskContext.get().taskAttemptId()
-  private val taskCPUs = TaskContext.get().cpus()
+  private val taskCPUs = TaskContext.get().cpus().toLong
   private val cometTaskMemoryManager = new CometTaskMemoryManager(id, taskAttemptId)
 
   private val plan = {
